+++ gliomoda/constants.py
