+++ test.py
@@ -1,0 +1,3 @@
+from gliomoda.model_handler import ModelHandler
+
+model_handler = ModelHandler(device="cuda")
